requirements.txt
# StegoVault - Python Dependencies
# Install with: pip install -r requirements.txt

# Image processing
Pillow>=9.0.0

# Cryptography for encryption/decryption
cryptography>=3.4.8

# GUI framework
# tkinter - built-in with most Python installations